Add simple ECS

+++ base/primitives.h
@@ -1,12 +1,14 @@
 #ifndef ADJUSTABLEPHYSICS2D_PRIMITIVES_H
 #define ADJUSTABLEPHYSICS2D_PRIMITIVES_H
 
+#include <cstddef>
+
 #ifdef DOUBLE_PRECISION
 using real = double;
 #else
 using real = float;
 #endif
 
-using EntityId = int;
+using EntityId = size_t;
 
 #endif //ADJUSTABLEPHYSICS2D_PRIMITIVES_H
